+++ src/main/java/roomescape/reservation/controller/dto/request/AdminCreateReservationRequest.java
@@ -2,9 +2,11 @@ package roomescape.reservation.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
-import roomescape.reservation.application.dto.request.CreateReservationWaitingServiceRequest;
+import roomescape.reservation.application.dto.request.CreateReservationServiceRequest;
 
-public record CreateReservationWaitingUserRequest(
+public record AdminCreateReservationRequest(
+        @NotNull(message = "예약자를 필수로 입력해야 합니다.")
+        Long memberId,
         @NotNull(message = "날짜를 필수로 입력해야 합니다.")
         LocalDate date,
         @NotNull(message = "시간을 필수로 입력해야 합니다.")
@@ -13,7 +15,7 @@ public record CreateReservationWaitingUserRequest(
         Long themeId
 ) {
 
-    public CreateReservationWaitingServiceRequest toServiceRequest(Long memberId) {
-        return new CreateReservationWaitingServiceRequest(memberId, date, timeId, themeId);
+    public CreateReservationServiceRequest toServiceRequest() {
+        return new CreateReservationServiceRequest(memberId, date, timeId, themeId);
     }
 }
