+++ src/main/java/roomescape/reservation/controller/dto/request/UserCreateReservationRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import roomescape.reservation.application.dto.request.CreateReservationServiceRequest;
 
-public record CreateReservationUserRequest(
+public record UserCreateReservationRequest(
         @NotNull(message = "날짜를 필수로 입력해야 합니다.")
         LocalDate date,
         @NotNull(message = "시간을 필수로 입력해야 합니다.")
