+++ src/main/java/roomescape/reservation/controller/dto/request/UserCreateReservationWaitingRequest.java
@@ -2,11 +2,9 @@ package roomescape.reservation.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
-import roomescape.reservation.application.dto.request.CreateReservationServiceRequest;
+import roomescape.reservation.application.dto.request.CreateReservationWaitingServiceRequest;
 
-public record CreateReservationAdminRequest(
-        @NotNull(message = "예약자를 필수로 입력해야 합니다.")
-        Long memberId,
+public record UserCreateReservationWaitingRequest(
         @NotNull(message = "날짜를 필수로 입력해야 합니다.")
         LocalDate date,
         @NotNull(message = "시간을 필수로 입력해야 합니다.")
@@ -15,7 +13,7 @@ public record CreateReservationAdminRequest(
         Long themeId
 ) {
 
-    public CreateReservationServiceRequest toServiceRequest() {
-        return new CreateReservationServiceRequest(memberId, date, timeId, themeId);
+    public CreateReservationWaitingServiceRequest toServiceRequest(Long memberId) {
+        return new CreateReservationWaitingServiceRequest(memberId, date, timeId, themeId);
     }
 }
